style: Add favicon
I created the favicon in Inkscape based on the Icon 16x16 template. For such small images, the stroke around the edges of the circle appears as individual stripes. To fix that, I set the stroke to a width of 0. I exported the image from Inkscape as a portable network graphic (PNG).

I installed the imagemagick package on Ubuntu, then used the `convert` command that it provides to turn the PNG into a favicon.

    convert favicon.png favicon.ico

Relates to #12 .